+++ example/react-example/.ddotrc.ts
@@ -1,11 +1,31 @@
 import { join } from 'path';
 export default {
   plugins: [
-    'react',
+    './test-plugin',
+    [
+      'jenkins',
+      {
+        hostName: 'jenkins.hello.com',
+        pathPrefix: '/job/job',
+        prompt: ({ branch, jobName }) => [
+          {
+            type: 'list',
+            name: 'Branch',
+            messages: 'which branch to deploy',
+            choices: () => [branch, 'develop', 'master'],
+          },
+          {
+            type: 'list',
+            name: 'Env',
+            messages: 'which Env to deploy',
+            choices: () => ['develop', 'release'],
+          },
+        ],
+      },
+    ],
     [
       'webpack',
       {
-        enableDll: false,
         proxy: {
           '/todos/1': {
             target: 'http://jsonplaceholder.typicode.com/',
@@ -22,6 +42,7 @@ export default {
             .entry('index')
             .add('./src/index')
             .end();
+
         },
       },
     ],