+++ packages/cli/src/__test__/index.test.ts
@@ -1,5 +1,0 @@
-import { getCommandName } from '../utils';
-
-test('getCommandName', () => {
-  expect(getCommandName('jenkins [aaa]')).toBe('jenkins');
-});